third_party/blink/web_tests/fast/dom/shadow/shadowroot-clonenode-expected.txt

Calling ShadowRoot.cloneNode() should throw a NotSupportedError exception.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS shadowRoot.cloneNode() threw exception NotSupportedError: Failed to execute 'cloneNode' on 'Node': ShadowRoot nodes are not clonable..
PASS successfullyParsed is true

TEST COMPLETE
